+++ consumer/consumertest/doc.go
@@ -12,21 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package translator
-
-import "go.opentelemetry.io/collector/consumer/pdata"
-
-type MetricsDecoder interface {
-	// DecodeMetrics converts a protocol-specific data model into pdata.
-	DecodeMetrics(src interface{}) (pdata.Metrics, error)
-}
-
-type TracesDecoder interface {
-	// DecodeTraces converts a protocol-specific data model into pdata.
-	DecodeTraces(src interface{}) (pdata.Traces, error)
-}
-
-type LogsDecoder interface {
-	// DecodeLogs converts a protocol-specific data model into pdata.
-	DecodeLogs(src interface{}) (pdata.Logs, error)
-}
+// Package consumertest defines types and functions used to help test packages
+// implementing the consumer package interfaces.
+package consumertest