+++ exporter/prometheusremotewriteexporter/exporter_test.go
@@ -30,7 +30,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/config/confighttp"
 	"go.opentelemetry.io/collector/consumer/pdata"
@@ -122,8 +121,7 @@ func Test_NewPRWExporter(t *testing.T) {
 				assert.Error(t, err)
 				return
 			}
-			require.NoError(t, err)
-			require.NoError(t, prwe.Start(context.Background(), componenttest.NewNopHost()))
+			require.NotNil(t, prwe)
 			assert.NotNil(t, prwe.namespace)
 			assert.NotNil(t, prwe.endpointURL)
 			assert.NotNil(t, prwe.externalLabels)
@@ -268,13 +266,6 @@ func runExportPipeline(ts *prompb.TimeSeries, endpoint *url.URL) []error {
 		errs = append(errs, err)
 		return errs
 	}
-
-	err = prwe.Start(context.Background(), componenttest.NewNopHost())
-	if err != nil {
-		errs = append(errs, err)
-		return errs
-	}
-
 	errs = append(errs, prwe.export(context.Background(), testmap)...)
 	return errs
 }
@@ -515,16 +506,15 @@ func Test_PushMetrics(t *testing.T) {
 				},
 				RemoteWriteQueue: RemoteWriteQueue{NumConsumers: 5},
 			}
-
 			assert.NotNil(t, cfg)
+			// c, err := config.HTTPClientSettings.ToClient()
+			// assert.Nil(t, err)
 			buildInfo := component.BuildInfo{
 				Description: "OpenTelemetry Collector",
 				Version:     "1.0",
 			}
 			prwe, nErr := NewPRWExporter(cfg, buildInfo)
 			require.NoError(t, nErr)
-			require.NoError(t, prwe.Start(context.Background(), componenttest.NewNopHost()))
-
 			err := prwe.PushMetrics(context.Background(), *tt.md)
 			if tt.returnErr {
 				assert.Error(t, err)